Add hover-expand sidebar, wire top-bar date range into more pages, fix forecast rounding
- New reusable Sidebar component: collapses to icons by default, expands on
  hover/focus, with a pin-open override in the top bar.
- The top-bar date-range dropdown now drives Transactions, History, and
  Analytics the same way it already did Overview.
- Fixed ai_sales_forecast(): regr_slope/regr_intercept return double
  precision regardless of input types, which made round(x, n) fail with
  "function round(double precision, integer) does not exist" -- cast back
  to numeric where they're computed.

Barcode label printing work (layout picker, thermal sizing) is intentionally
left out of this commit.

+++ src/datatabase/pharmacy_schema_consolidated.sql
@@ -4968,8 +4968,12 @@ begin
   stats as (
     select
       coalesce(avg(y), 0) as avg_qty,
-      coalesce(regr_slope(y, x), 0) as slope,
-      coalesce(regr_intercept(y, x), avg(y), 0) as intercept,
+      -- regr_slope/regr_intercept always return double precision in Postgres,
+      -- regardless of the input types (x/y are already cast to numeric above)
+      -- -- cast back to numeric here so every round(x, n) below resolves to
+      -- round(numeric, integer); round(double precision, integer) doesn't exist.
+      coalesce(regr_slope(y, x), 0)::numeric as slope,
+      coalesce(regr_intercept(y, x), avg(y), 0)::numeric as intercept,
       coalesce(sum(revenue) / nullif(sum(y), 0), 0) as avg_unit_revenue,
       coalesce(max(x), 0) as max_x
     from numbered
